Adding changes to complete the order form

+++ src/carts/views.py
@@ -1,11 +1,14 @@
+from django.contrib.auth.forms import AuthenticationForm
+from django.core.urlresolvers import reverse
 from django.views.generic.base import View
-from django.views.generic.detail import SingleObjectMixin
-from django.http import HttpResponseRedirect,Http404
-from django.shortcuts import render,get_object_or_404
+from django.views.generic.detail import SingleObjectMixin,DetailView
+from django.http import HttpResponseRedirect,Http404,JsonResponse
+from django.shortcuts import render,get_object_or_404,redirect
+from django.views.generic.edit import FormMixin
 
 from products.models import Variation
 from .models import Cart, CartItem
-
+from orders.forms import GuestCheckoutForm
 # Create your views here.
 
 class CartView(SingleObjectMixin,View):
@@ -30,7 +33,8 @@ class CartView(SingleObjectMixin,View):
     def get(self,request,*args,**kwargs):
         cart = self.get_object()
         item_id = request.GET.get("item")
-        delete_item = request.GET.get("delete")
+        delete_item = request.GET.get("delete",False) # Default for deleted items will be set to False.
+        item_added = False # Default for added items
         if item_id:
             item_instance = get_object_or_404(Variation,id = item_id)
             qty = request.GET.get("qty",1) #Giving the quantity a default value of 1
@@ -39,16 +43,72 @@ class CartView(SingleObjectMixin,View):
                     delete_item = True
             except:
                 raise Http404
-            cart_item = CartItem.objects.get_or_create(cart=cart, item = item_instance)[0]
+            cart_item,created_item= CartItem.objects.get_or_create(cart=cart, item = item_instance) # We will either have a created item, or an item that is already in the cart
+            if created_item:
+                item_added = True # If a new item is added to the cart it will return as True
             if delete_item:
                 cart_item.delete()
             else:
                 cart_item.quantity = qty
                 cart_item.save()
-                print(cart_item)
+            if not request.is_ajax():
+                return HttpResponseRedirect(reverse("cart"))
+        if request.is_ajax(): # If an AJAX call happens then...
+            print(request.GET.get("item")) # Then print out what the request gives back
+            try:
+                total = cart_item.line_item_total
+            except:
+                total = None
+            try:
+                subtotal = cart_item.cart_subtotal
+            except:
+                subtotal = None
+            data = {
+                "deleted": delete_item,
+                "item_added": item_added,
+                "line_total": total,
+                "subtotal":subtotal,
+            }
+            return JsonResponse(data) # Return a JSON Response.
         context = {
             "object":self.get_object()
         }
-        print(context)
         template = self.template_name
         return render(request,template,context)
+
+
+
+class CheckoutView(FormMixin, DetailView):
+	model = Cart
+	template_name = "carts/checkout_view.html"
+	form_class = GuestCheckoutForm
+
+	def get_object(self, *args, **kwargs):
+		cart_id = self.request.session.get("cart_id")
+		if cart_id == None:
+			return redirect("cart")
+		cart = Cart.objects.get(id=cart_id)
+		return cart
+
+	def get_context_data(self, *args, **kwargs):
+		context = super(CheckoutView, self).get_context_data(*args, **kwargs)
+		user_can_continue = False
+		if not self.request.user.is_authenticated():# or if request.user.is_guest:
+			context["login_form"] = AuthenticationForm()
+			context["next_url"] = self.request.build_absolute_uri()
+		if self.request.user.is_authenticated(): #or if request.user.is_guest():
+			user_can_continue = True
+		context["user_can_continue"] = user_can_continue
+		context["form"] = self.get_form()
+		return context
+
+	def post(self, request, *args, **kwargs):
+		form = self.get_form()
+		if form.is_valid():
+			print(form.cleaned_data.get("email"))
+			return self.form_valid(form)
+		else:
+			return self.form_invalid(form)
+
+	def get_success_url(self):
+		return reverse("checkout")
